refactor(routes): extract reaction routes to lib/routes/reactions.js
Move the three reaction API endpoints out of server.js into
lib/routes/reactions.js as a factory function accepting deps
(isAuthenticated, requireSessionAccess, authMode, reactions).

- GET /api/reactions/:sessionKey
- GET /api/messages/:messageId/reactions
- POST /api/messages/:messageId/reactions

Reduces server.js by ~47 lines. Keeps middleware order and error
contracts identical.
